pulled out the functions into a separate module + docs.
