+++ skit_pipelines/components/gen_eer_metrics.py
@@ -27,6 +27,6 @@ def main(data_path):
 
 # if __name__ == "__main__":
 
-    # main("duck_4284.csv")
+#     main("duck_4284.csv")
     # main("duck_4333.csv")
