+++ skit_pipelines/components/modify_entities/__init__.py
@@ -1,0 +1,3 @@
+
+from skit_pipelines.components.modify_entities.duckling_inference import modify_truth
+from skit_pipelines.components.modify_entities.modify_predictions import modify_predictions
