+++ source/index.rst
@@ -18,6 +18,7 @@ skit-pipelines
    Train XLMR model <./skit_pipelines.pipelines.train_voicebot_intent_model_xlmr>
    XLMR Intent evaluation <./skit_pipelines.pipelines.eval_voicebot_xlmr_pipeline>
    Annotated Dataset Intent Evaluation <./skit_pipelines.pipelines.irr_from_tog>
+   Download tagged entity dataset <./skit_pipelines.pipelines.fetch_tagged_entity_dataset>
 
 .. toctree::
    :maxdepth: 1
@@ -77,6 +78,8 @@ Here's a list of official pipelines, within these docs we share snippets for sla
 | 6 | :ref:`XLMR Intent evaluation <p_eval_voicebot_xlmr_pipeline>`                   |
 +--+----------------------------------------------------------------------------------+
 | 7 | :ref:`Annotated Dataset Intent Evaluation <p_irr_from_tog>`                     |
++--+----------------------------------------------------------------------------------+
+| 8 | :ref:`Download tagged entity dataset <p_fetch_tagged_entity_dataset>`           |
 +--+----------------------------------------------------------------------------------+
 
 Project strucuture
